GH-406: Move extension registration helper back to n4jscbase bundle
