--- plugins/org.eclipse.n4js.headless.helper/.gitignore
@@ -1,1 +1,0 @@
-/src-gen
