--- plugins/org.eclipse.n4js.headless.helper/.settings/org.eclipse.core.runtime.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-line.separator=\n
