--- plugins/org.eclipse.n4js.headless.helper/.settings/org.eclipse.emf.ecore.xcore.Xcore.prefs
@@ -1,10 +1,0 @@
-autobuilding=true
-eclipse.preferences.version=1
-is_project_specific=true
-outlet.DEFAULT_OUTPUT.cleanDirectory=true
-outlet.DEFAULT_OUTPUT.cleanupDerived=true
-outlet.DEFAULT_OUTPUT.createDirectory=true
-outlet.DEFAULT_OUTPUT.derived=true
-outlet.DEFAULT_OUTPUT.hideLocalSyntheticVariables=true
-outlet.DEFAULT_OUTPUT.installDslAsPrimarySource=false
-outlet.DEFAULT_OUTPUT.override=true
