--- plugins/org.eclipse.n4js.headless.helper/.settings/org.eclipse.jdt.launching.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-org.eclipse.jdt.launching.PREF_STRICTLY_COMPATIBLE_JRE_NOT_AVAILABLE=warning
