--- plugins/org.eclipse.n4js.headless.helper/.settings/org.eclipse.ltk.core.refactoring.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-org.eclipse.ltk.core.refactoring.enable.project.refactoring.history=false
