--- plugins/org.eclipse.n4js.headless.helper/.settings/org.eclipse.xtend.core.Xtend.prefs
@@ -1,39 +1,0 @@
-.version=1
-autobuilding=true
-blank.line.around.expressions=0
-blank.lines.after.imports=1
-blank.lines.after.last.member=0
-blank.lines.after.packagedecl=1
-blank.lines.before.first.member=0
-blank.lines.between.classes=1
-blank.lines.between.enum.literals=0
-blank.lines.between.fields=0
-blank.lines.between.fields.and.methods=1
-blank.lines.between.imports=0
-blank.lines.between.methods=1
-braces.in.new.line=false
-eclipse.preferences.version=1
-indentation=\t
-indentation.length=4
-is_project_specific=true
-line.separator=\n
-line.width.max=120
-newline.after.class.annotations=false
-newline.after.constructor.annotations=false
-newline.after.field.annotations=false
-newline.after.method.annotations=false
-newline.after.parameter.annotations=false
-org.eclipse.xtend.core.validation.IssueCodes.java_doc_linking=warning
-org.eclipse.xtend.ide.default.eclipse_profile=_N4JS
-org.eclipse.xtext.xbase.validation.IssueCodes.equals_with_null=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.unhandled_exception=warning
-outlet.DEFAULT_OUTPUT.cleanDirectory=false
-outlet.DEFAULT_OUTPUT.cleanupDerived=true
-outlet.DEFAULT_OUTPUT.createDirectory=true
-outlet.DEFAULT_OUTPUT.derived=true
-outlet.DEFAULT_OUTPUT.hideLocalSyntheticVariables=true
-outlet.DEFAULT_OUTPUT.installDslAsPrimarySource=false
-outlet.DEFAULT_OUTPUT.override=true
-preserve.blank.lines=1
-preserve.new.lines=true
-whitespace.between.keyword.and.parenthesis=true
