--- plugins/org.eclipse.n4js.headless.helper/META-INF/MANIFEST.MF
@@ -1,17 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: %pluginName
-Bundle-SymbolicName: org.eclipse.n4js.headless.helper
-Bundle-Version: 0.0.1.qualifier
-Bundle-Vendor: %providerName
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Require-Bundle: org.eclipse.n4js,
- org.eclipse.n4js.utils,
- org.eclipse.n4js.transpiler.es,
- org.eclipse.n4js.runner,
- org.eclipse.n4js.tester,
- org.eclipse.n4js.runner.nodejs,
- org.eclipse.n4js.tester.nodejs,
- org.eclipse.core.runtime
-Export-Package: org.eclipse.n4js.headless.helper
-
