--- plugins/org.eclipse.n4js.headless.helper/plugin.properties
@@ -1,3 +1,0 @@
-
-pluginName = org.eclipse.n4js.generator.headless
-providerName = Eclipse N4JS Project
