+++ plugins/pom.xml
@@ -82,8 +82,6 @@ Contributors:
 		<!-- npm exporter -->
 		<module>org.eclipse.n4js.npmexporter</module>
 		<module>org.eclipse.n4js.npmexporter.ui</module>
-		<!-- headless helper -->
-		<module>org.eclipse.n4js.headless.helper</module>
 	</modules>
 
 	<properties>
