+++ testhelpers/org.eclipse.n4js.tests.helper/META-INF/MANIFEST.MF
@@ -23,8 +23,7 @@ Require-Bundle: org.eclipse.n4js,
  org.eclipse.n4js.runner,
  org.eclipse.n4js.runner.nodejs,
  org.eclipse.n4js.generator.headless,
- org.eclipse.n4js.hlc.base,
- org.eclipse.n4js.headless.helper
+ org.eclipse.n4js.hlc.base
 Export-Package: org.eclipse.n4js,
  org.eclipse.n4js.analysis,
  org.eclipse.n4js.csv,
