+++ tests/org.eclipse.n4js.accesscontrol.tests/META-INF/MANIFEST.MF
@@ -19,5 +19,6 @@ Require-Bundle: com.google.inject,
  org.eclipse.n4js.ts,
  org.eclipse.n4js.n4mf.model,
  org.eclipse.n4js.utils,
- org.eclipse.n4js.tests.helper
+ org.eclipse.n4js.tests.helper,
+ org.eclipse.n4js.hlc.base
