+++ tests/org.eclipse.n4js.generator.headless.tests/META-INF/MANIFEST.MF
@@ -20,5 +20,6 @@ Require-Bundle: com.google.inject,
  org.eclipse.n4js.n4mf.model,
  org.eclipse.n4js.utils,
  org.eclipse.n4js.tests.helper,
- org.eclipse.n4js.transpiler.es
+ org.eclipse.n4js.transpiler.es,
+ org.eclipse.n4js.hlc.base
