+++ tests/org.eclipse.n4js.tester.tests/META-INF/MANIFEST.MF
@@ -9,4 +9,4 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: org.eclipse.n4js.tester.tests.helper,
  org.eclipse.xtend.lib,
  org.junit,
- org.eclipse.n4js.headless.helper
+ org.eclipse.n4js.hlc.base
