+++ tools/org.eclipse.n4js.hlc.base/META-INF/MANIFEST.MF
@@ -16,6 +16,8 @@ Require-Bundle: org.eclipse.n4js,
  org.eclipse.n4js.tester,
  org.eclipse.n4js.external.libraries,
  org.eclipse.core.runtime,
- org.eclipse.n4js.headless.helper
+ org.eclipse.n4js.transpiler.es,
+ org.eclipse.n4js.runner.nodejs,
+ org.eclipse.n4js.tester.nodejs
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.n4js.hlc.base
