+++ tools/org.eclipse.n4js.hlc.base/src/org/eclipse/n4js/hlc/base/HeadlessExtensionRegistrationHelper.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.headless.helper;
+package org.eclipse.n4js.hlc.base;
 
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.n4js.N4JSGlobals;
